a: 添加travis 测试 f: 修改 Gethostbyaddr 函数 函数中 names[0] 使用前未对 names长度进行判断 m: 修改 IsFile 函数    函数判断文件是否存在的同时还需判断是否不为目录文件 m: 拆分模块 分为不同文件 m: 拆分测试模块,支持windows测试 m: 修改 Pack 和 Unpack 的测试用例 去除输出的内容 改为标准测试
