+++ .travis.yml
@@ -1,4 +1,8 @@
 language: go
 
 go:
-- 1.9.x
+	- 1.9.x
+	- 1.12.x
+	- 1.13.x
+script:
+	- go test -v ./...
