+++ network.go
@@ -1,0 +1,75 @@
+package php2go
+
+import (
+	"encoding/binary"
+	"net"
+	"os"
+	"strings"
+)
+
+// ////////// Network Functions ////////////
+
+// Gethostname gethostname()
+func Gethostname() (string, error) {
+	return os.Hostname()
+}
+
+// Gethostbyname gethostbyname()
+// Get the IPv4 address corresponding to a given Internet host name
+func Gethostbyname(hostname string) (string, error) {
+	ips, err := net.LookupIP(hostname)
+	if ips != nil {
+		for _, v := range ips {
+			if v.To4() != nil {
+				return v.String(), nil
+			}
+		}
+		return "", nil
+	}
+	return "", err
+}
+
+// Gethostbynamel gethostbynamel()
+// Get a list of IPv4 addresses corresponding to a given Internet host name
+func Gethostbynamel(hostname string) ([]string, error) {
+	ips, err := net.LookupIP(hostname)
+	if ips != nil {
+		var ipstrs []string
+		for _, v := range ips {
+			if v.To4() != nil {
+				ipstrs = append(ipstrs, v.String())
+			}
+		}
+		return ipstrs, nil
+	}
+	return nil, err
+}
+
+// Gethostbyaddr gethostbyaddr()
+// Get the Internet host name corresponding to a given IP address
+func Gethostbyaddr(ipAddress string) (string, error) {
+	names, err := net.LookupAddr(ipAddress)
+	if names != nil && len(names) != 0 {
+		return strings.TrimRight(names[0], "."), nil
+	}
+	return "", err
+}
+
+// IP2long ip2long()
+// IPv4
+func IP2long(ipAddress string) uint32 {
+	ip := net.ParseIP(ipAddress)
+	if ip == nil {
+		return 0
+	}
+	return binary.BigEndian.Uint32(ip.To4())
+}
+
+// Long2ip long2ip()
+// IPv4
+func Long2ip(properAddress uint32) string {
+	ipByte := make([]byte, 4)
+	binary.BigEndian.PutUint32(ipByte, properAddress)
+	ip := net.IP(ipByte)
+	return ip.String()
+}
