--- php_windows.go
@@ -1,5 +1,0 @@
-// php2go functions
-
-// +build windowns
-
-package php2go
